# feat(wal): add perf in command line tools