… .golangci.yaml