Add a command to test wal's performance, usage:

```
go run cmd/main.go wal perf --shard 0 --wal-dir test --namespace test --reader-count 3
```

Output:

```
Thread: Writer, Total entry: 1000000, Total time: 42.770528s, Throughput: 23380.585809 ops/s
Thread: Reader0, Total entry: 1000000, Total time: 42.770618s, Throughput: 23380.536610 ops/s
Thread: Reader1, Total entry: 1000000, Total time: 42.770535s, Throughput: 23380.581982 ops/s
Thread: Reader2, Total entry: 1000000, Total time: 42.77062s, Throughput: 23380.535517 ops/s

Histogram of Writer latency:
P50:  41µs
P90:  47µs 
P99:  63µs 
P999: 215µs

Histogram of Reader latency:
P50:  45µs
P90:  53µs 
P99:  70µs 
P999: 220µs
```